requirements.txt
pandas~=2.2.3
langchain-core~=0.3.25
numpy~=1.26.4
rouge-score~=0.1.2
sentence-transformers~=3.2.0
transformers~=4.45.2
langchain-community~=0.3.12
langchain-text-splitters~=0.3.3
torch~=2.5.0
python-dotenv~=1.0.1